feat(docs): Include aztec rpc interface typedoc output in docs
