+++ docs/package.json
@@ -38,7 +38,10 @@
     "@docusaurus/module-type-aliases": "^2.4.1",
     "@tsconfig/docusaurus": "^1.0.5",
     "concurrently": "^8.0.1",
+    "docusaurus-plugin-typedoc": "^0.20.1",
     "nodemon": "^3.0.1",
+    "typedoc": "^0.25.1",
+    "typedoc-plugin-markdown": "^3.16.0",
     "typescript": "^4.7.2"
   },
   "browserslist": {
